AnyCartProject-master/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
